Add a new function and edit the playRound function
The new function uppers the first letter of a string, to display the
first letters of rock, paper and scissor in upper case.
Edited the playRound function to display more info to the player, for
example: "You Win! Rock beats Scissor." instead of "Win".
Add new line to actually start one round.