Add new function "playGame" to play a "match"
The new function "playGame" declares a human- and a computer-score and
starts the game in a loop. The loop ends, when the computer or the
player reaches 3 wins in total. The score will be displayed to the
player via console.
Edited the "playRound" function, it now returns "win" or "lose", so
that the "playGame" function can count a score.